+++ packages/core/client/src/schema-component/antd/error-fallback/ErrorFallback.tsx
@@ -19,6 +19,7 @@ import { useLocation } from 'react-router-dom';
 const { Paragraph, Text, Link } = Typography;
 
 export const useDownloadLogs = (error: any, data: Record<string, any> = {}) => {
+  const location = useLocation();
   const [loading, setLoading] = React.useState(false);
   const api = useAPIClient();
   return {
@@ -35,6 +36,7 @@ export const useDownloadLogs = (error: any, data: Record<string, any> = {}) => {
               message: error.message,
               stack: error.stack,
             },
+            location,
             ...data,
           },
         });
@@ -56,9 +58,8 @@ export const ErrorFallback: FC<FallbackProps> & {
   Modal: FC<FallbackProps>;
 } = ({ error }) => {
   const schema = useFieldSchema();
-  const location = useLocation();
   const { t } = useTranslation();
-  const { loading, download } = useDownloadLogs(error, { schema, location });
+  const { loading, download } = useDownloadLogs(error, { schema });
 
   const subTitle = (
     <Trans>
