--- packages/core/client/src/schema-component/antd/error-fallback/__tests__/components/basic.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { ErrorBoundary } from 'react-error-boundary';
-import { ErrorFallback } from '../../ErrorFallback';
-
-const App = () => {
-  throw new Error('error message');
-};
-
-export default () => {
-  return (
-    <ErrorBoundary FallbackComponent={ErrorFallback} onError={console.error}>
-      <App />
-    </ErrorBoundary>
-  );
-};
