+++ packages/core/client/src/schema-component/antd/error-fallback/index.en-US.md
@@ -6,8 +6,8 @@ Based on the [react-error-boundary](https://github.com/bvaughn/react-error-bound
 
 ## Basic
 
-<code src="./demos/new-demos/basic.tsx"></code>
+<code src="./demos/basic.tsx"></code>
 
 ## Modal
 
-<code src="./demos/new-demos/modal.tsx"></code>
+<code src="./demos/modal.tsx"></code>
